+++ test/Braintree.Tests/TransactionCreditCardRequestTest.cs
@@ -12,5 +12,19 @@ namespace Braintree.Tests
             request.Token = "my-token";
             Assert.IsTrue(request.ToXml().Contains("my-token"));
         }
+
+        [Test]
+        public void ToXml_Includes_NetworkTokenizationAttributesRequest()
+        {
+            TransactionCreditCardRequest request = new TransactionCreditCardRequest();
+            request.NetworkTokenizationAttributes = new NetworkTokenizationAttributesRequest();
+            request.NetworkTokenizationAttributes.Cryptogram = "validcryptogram";
+            request.NetworkTokenizationAttributes.EcommerceIndicator = "05";
+            request.NetworkTokenizationAttributes.TokenRequestorId = "123456";
+
+            Assert.IsTrue(request.ToXml().Contains("validcryptogram"));
+            Assert.IsTrue(request.ToXml().Contains("05"));
+            Assert.IsTrue(request.ToXml().Contains("123456"));
+        }
     }
 }